Edit comments. Fix methods to work more in a more logical way.

+++ Database_Project_Group1/src/Artist.java
@@ -47,49 +47,64 @@ public class Artist {
     }
 
     /**
-     * selectAll method of the Artist class.
+     * selectArtist method of the Artist class.
      * This method will use the getData(String, ArrayList<String>) method of the Alinity
-     * class. The SQL statement will select Albums where the ID value input will be bound
+     * class. The SQL statement will select Albums where the name of the artist input will be bound
      * from the ArrayList, and save the result.
      * Sets ID, name, info, recordLabelId based on result data.
      * This method will specifically work only when selecting ALL the information for said Artist
-     * id.
+     * name.
+     * Both admins and general users will have access to selecting artists from search.
      * If the attempt of the executing the statement fails, log the error to the file.
      *
      * @throws AlinityException
      */
-    public void selectAll(ArrayList<String> stringList) throws AlinityException {
+    public void selectArtist(User user, String artistName) throws AlinityException {
         try {
-            ArrayList<ArrayList<String>> result = AlinityMain.alinityDB.getData("SELECT * FROM Artist WHERE artistId = ?" , stringList);
-            System.out.print("\nColumn headers: " + result.get(0));
-            ArrayList<String> artistData = result.get(1);
-            setArtistId(Integer.parseInt(artistData.get(0)));
-            setArtistName(artistData.get(1));
-            setArtistInfo(artistData.get(2));
-            setRecordLabelId(Integer.parseInt(artistData.get(3)));
+            if(user.getRole().equals("General") || user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(artistName);
+                ArrayList<ArrayList<String>> result = AlinityMain.alinityDB.getData("SELECT * FROM Artist WHERE artistName = ?" , info);
+                System.out.print("\nColumn headers: " + result.get(0));
+                ArrayList<String> artistData = result.get(1);
+                setArtistId(Integer.parseInt(artistData.get(0)));
+                setArtistName(artistData.get(1));
+                setArtistInfo(artistData.get(2));
+                setRecordLabelId(Integer.parseInt(artistData.get(3)));
+                printArtist();
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator.");
         } catch (IndexOutOfBoundsException ioobe) {
-            throw new AlinityException(ioobe, "-> Error in obtaining data (IndexOutOfBoundsException) from the database. Please check your syntax in the selectAll(ArrayList<String>) method.", "SELECT * FROM Artist WHERE artistId = ?");
+            throw new AlinityException(ioobe, "-> Error in obtaining data (IndexOutOfBoundsException) from the database. Please check your syntax in the selectAll(ArrayList<String>) method.", "SELECT * FROM Artist WHERE artistName = ?");
         } catch (NullPointerException npe) {
-            throw new AlinityException(npe, "-> Error in obtaining data (NullPointerException) from the database. Please check your syntax in the selectAll(ArrayList<String>) method.", "SELECT * FROM Artist WHERE artistId = ?");
+            throw new AlinityException(npe, "-> Error in obtaining data (NullPointerException) from the database. Please check your syntax in the selectAll(ArrayList<String>) method.", "SELECT * FROM Artist WHERE artistName = ?");
         }
     }
 
     /**
-     * updateAll method of the Artist class.
+     * updateArtist method of the Artist class.
      * A SQL statement as a String is created using the current name, info,
-     * and recordLabelId where the artistId is based on the values which will be bound
-     * from the input ArrayList of Strings.
+     * and recordLabelId. They are bound to the statement using an ArrayList of Strings.
+     * The artistId where these changes will take place is taken from the object
+     * on which this method will be called on.
      * Executes using the setDada method.
      * This method will specifically work only when updating ALL the information for said artist
      * id.
+     * Only works for admins, since users with the General role should not be allowed to update information.
      * If the attempt of the executing the statement fails, log the error to the file.
      *
      * @throws AlinityException
      */
-    public boolean updateAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean updateArtist(User user, String artistName, String artistInfo, int recordLabelId) throws AlinityException {
         try {
-            String putStmt = "UPDATE Artist SET artistName = ? , artistInfo = ?, recordLabelId = ? WHERE artistId = ?";
-            return AlinityMain.alinityDB.setData(putStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(artistName);
+                info.add(artistInfo);
+                info.add(String.valueOf(recordLabelId));
+                info.add(String.valueOf(this.getArtistId()));
+                String putStmt = "UPDATE Artist SET artistName = ? , artistInfo = ?, recordLabelId = ? WHERE artistId = ?";
+                return AlinityMain.alinityDB.setData(putStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
             throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the updateAll(ArrayList<String>) method.", "UPDATE Artist SET artistName = ? , artistInfo = ?, recordLabelId = ? WHERE artistId = ?");
         } catch (IndexOutOfBoundsException ioobe) {
@@ -98,19 +113,27 @@ public class Artist {
     }
 
     /**
-     * insertAll method of the Artist class.
+     * insertArtist method of the Artist class.
      * A SQL statement as a String is created, which takes in
      * values based on the values to be bound from the ArrayList of Strings.
      * Executes using the setData method.
      * This method will specifically work only when inserting ALL the information for said artist.
+     * Only admins have the permissions to do this. General users should not be allowed to insert
+     * new information to the database.
      * If the attempt of the executing the statement fails, log the error to the file.
      *
      * @throws AlinityException
      */
-    public boolean insertAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean insertArtist(User user, String artistName, String artistInfo, int recordLabelId) throws AlinityException {
         try {
-            String insertStmt = "INSERT INTO Artist (artistName, artistInfo, recordLabelId) VALUES (?, ?, ?)";
-            return AlinityMain.alinityDB.setData(insertStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(artistName);
+                info.add(artistInfo);
+                info.add(String.valueOf(recordLabelId));
+                String insertStmt = "INSERT INTO Artist (artistName, artistInfo, recordLabelId) VALUES (?, ?, ?)";
+                return AlinityMain.alinityDB.setData(insertStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
             throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the insertAll(ArrayList<String>) method.","INSERT INTO Artist (artistName, artistInfo, recordLabelId) VALUES (?, ?, ?)");
         } catch (IndexOutOfBoundsException ioobe) {
@@ -119,7 +142,7 @@ public class Artist {
     }
 
     /**
-     * deleteAll method of the Artist class.
+     * deleteArtist method of the Artist class.
      * Deletes the given artist data where the artistId is bound
      * via tha value inside the ArrayList of Strings.
      * Executes using the setData method.
@@ -129,10 +152,14 @@ public class Artist {
      *
      * @throws AlinityException
      */
-    public boolean deleteAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean deleteArtist(User user) throws AlinityException {
         try {
-            String deleteStmt = "DELETE FROM Artist WHERE artistId = ?";
-            return AlinityMain.alinityDB.setData(deleteStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(String.valueOf(this.getArtistId()));
+                String deleteStmt = "DELETE FROM Artist WHERE artistId = ?";
+                return AlinityMain.alinityDB.setData(deleteStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
             throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the deleteAll() method.", "DELETE FROM Artist WHERE artistId = ?");
         } catch (IndexOutOfBoundsException ioobe) {
